+++ changelog.md
@@ -4,6 +4,7 @@
 	- count `gifs`, `hello` words, `hungry` words, and `emojis` for accepting users
 	- count `emotes` occurences and `cats` pictures sent for whole guild
 - Feat: `dbReturnType` frozen object with returned values for db helpers ([#234](https://github.com/Eccleria/ewibot/pull/234))
+- Fix: alavirien wrong db access when alavirien check ([#250](https://github.com/Eccleria/ewibot/pull/234))
 
 ## 5.4.0
 - Feat: add reaction if is mentioned by a user ([#230](https://github.com/Eccleria/ewibot/pull/230))